IGNITE-24847 Move GridFunc methods
